ugly testing

+++ energyAPI.py
@@ -17,7 +17,7 @@ def root():
 @app.get("/latest/")
 def return_metrics(metric: Optional[str] = None, n: Optional[int] = 1):
     data = {}
-    lines = tail(f=f'{os.path.abspath(os.path.dirname(__file__))}/P1_log.csv', n=21*n).reverse()
+    lines = tail(f=f'{os.path.abspath(os.path.dirname(__file__))}/P1_log.csv', n=int(21*n)).reverse()
 
     print(lines)
     # last_sample = lines[0].split(",")[1]
